refactor(drawing-2d): split the paper transform out of svg-exporter.ts
The `svgNum()` guard and the `computeTransform` bounds sanitising added by
this PR took `svg-exporter.ts` to 652 lines against a budget of 613, failing
the module-size ratchet. Split rather than raise.

`computeTransform`, `scaleLabel` and `transformPoint` move to a new
`svg-transform.ts` alongside `Transform2D` and `finiteOr0`. Fitting a
drawing's world bounds onto a sheet — and reporting the scale actually used —
is one cohesive job, separable from turning the fitted geometry into markup.
None of the three ever read `this`, so they become plain functions and the
call sites drop the `this.` prefix.

`createTitleBlock`'s `scaleLabel` parameter is renamed `scaleText`: that name
now belongs to the imported function, and shadowing it there would read as a
call site.

A pure move, proven rather than asserted: the moved bodies `diff` clean
against the originals modulo the dedent and the `private` -> `export
function` keyword swap, and the `Transform2D` body is byte-identical. Full
`packages/drawing-2d` suite green (518 tests, 39 files), as is
`packages/export`.

The `svg-exporter.ts` row ratchets DOWN 613 -> 537 to the new measured count,
with `packages/drawing-2d`'s ALLOWLIST_DIGESTS entry re-pinned in the same
commit. Edited by hand rather than with `--update`, which would also have
ratcheted the unrelated headroom rows the gate notes for
`packages/export/src/schema-converter.ts` and
`packages/server-client/src/parquet-tables.ts` — those belong to other PRs.
No row moved up.
